move chain fixture to agent0.test_fixtures
