+++ conftest.py
@@ -9,6 +9,7 @@ import pytest
 from chainsync.test_fixtures import database_engine, db_api, db_session, dummy_session, psql_docker
 from ethpy.test_fixtures import hyperdrive_interface, init_local_hyperdrive_pool, local_chain, local_hyperdrive_pool
 
+from agent0.test_fixtures import chain
 from agent0.test_utils import cycle_trade_policy
 
 # Hack to allow for vscode debugger to throw exception immediately
@@ -50,6 +51,7 @@ if os.getenv("_PYTEST_RAISE", "0") != "0":
 # This allows for users of fixtures to not have to import all dependency fixtures when running
 # TODO this means pytest can only be ran from this directory
 __all__ = [
+    "chain",
     "database_engine",
     "db_api",
     "db_session",
